Fixed euler

+++ model/point.py
@@ -4,11 +4,11 @@ import model
 
 
 class Point:
-    def __init__(self, x: np.float = 0., y: np.float = 0., z: np.float = 0., m: np.float = 0.):
-        self.pos: np.ndarray = np.array([x, y, z])
-        self.vel: np.ndarray = np.array([0., 0., 0.])
-        self.acc: np.ndarray = np.array([0., 0., 0.])
-        self.mass: np.float = np.float(m)
+    def __init__(self, x: np.float128 = 0., y: np.float128 = 0., z: np.float128 = 0., m: np.float128 = 0.):
+        self.pos: np.ndarray = np.array([x, y, z], dtype=np.float128)
+        self.vel: np.ndarray = np.array([0., 0., 0.], dtype=np.float128)
+        self.acc: np.ndarray = np.array([0., 0., 0.], dtype=np.float128)
+        self.mass: np.float = np.float128(m)
 
     def __str__(self):
         x, y, z = self.pos
